🔄 refactor(config): Move `connectWithTimeout` Outside of Helpers Module (#1513)
