+++ config/delete-user.js
@@ -1,10 +1,11 @@
 const path = require('path');
 require('module-alias')({ base: path.resolve(__dirname, '..', 'api') });
-const { connectWithTimeout, askQuestion, silentExit } = require('./helpers');
+const { askQuestion, silentExit } = require('./helpers');
 const User = require('~/models/User');
+const connect = require('./connect');
 
 (async () => {
-  await connectWithTimeout();
+  await connect();
 
   /**
    * Show the welcome / help menu
