+++ config/helpers.js
@@ -6,8 +6,6 @@ const fs = require('fs');
 const path = require('path');
 const readline = require('readline');
 const { execSync } = require('child_process');
-require('module-alias')({ base: path.resolve(__dirname, '..', 'api') });
-const connectDb = require('~/lib/db/connectDb');
 
 const askQuestion = (query) => {
   const rl = readline.createInterface({
@@ -45,33 +43,6 @@ const silentExit = (code = 0) => {
   process.exit(code);
 };
 
-async function connectWithTimeout() {
-  /**
-   * Connect to the database
-   * - If it takes a while, we'll warn the user
-   */
-  let timeout = setTimeout(() => {
-    console.orange(
-      'This is taking a while... You may need to check your connection if this fails.',
-    );
-    timeout = setTimeout(() => {
-      console.orange('Still going... Might as well assume the connection failed...');
-      timeout = setTimeout(() => {
-        console.orange('Error incoming in 3... 2... 1...');
-      }, 13000);
-    }, 10000);
-  }, 5000);
-  // Attempt to connect to the database
-  try {
-    console.orange('Warming up the engines...');
-    await connectDb();
-    clearTimeout(timeout);
-  } catch (e) {
-    console.error(e);
-    silentExit(1);
-  }
-}
-
 // Set the console colours
 console.orange = (msg) => console.log('\x1b[33m%s\x1b[0m', msg);
 console.green = (msg) => console.log('\x1b[32m%s\x1b[0m', msg);
@@ -87,6 +58,5 @@ module.exports = {
   askQuestion,
   silentExit,
   isDockerRunning,
-  connectWithTimeout,
   deleteNodeModules,
 };
